pyproject.toml
[project]
name = "trading"
version = "0.1.0"
description = "ML-powered trading system for SPY with rigorous time-series validation."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "yfinance>=0.2",
    "pandas>=2.0",
    "pyarrow>=15.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/trading"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "ruff>=0.6",
    "pre-commit>=3.7",
    "jupyterlab>=4.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "D", "ANN"]
ignore = [
    "D203",  # incompatible with D211
    "D213",  # incompatible with D212
    "ANN101",  # self annotation (deprecated in newer ruff)
    "ANN102",  # cls annotation (deprecated in newer ruff)
    "ANN401",
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["D", "ANN"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra --strict-markers"
